Data Visualization Project for WGU assessment
This project is the final assessment I did to pass the Advanced Data Visualization course. I collected all the neccessary data on the website, and imported them to the Tableau desktop. I also created the final dashboard to tell the story about the reason I chose the 2019 Honda CR-V AWD as the best option for the company car purchasing.